moved progs to subdirs, run_tests.py is WIP

+++ run_tests.py
@@ -4,6 +4,7 @@
 import subprocess
 import os
 import argparse
+import re
 from build import compile_rc
 
 curr_path = os.path.dirname(os.path.realpath(__file__))
@@ -19,32 +20,65 @@ def unit_tests():
         curr_test_path = os.path.join(test_path, test)
         subprocess.run([curr_test_path])
 
-# compiles and runs everything in ../programs and compares the return code
-# to the output from the racket interpreter
-def full_tests():
-    prog_path = os.path.join(curr_path, 'programs')
-    prog_files = list(filter(lambda x: x.endswith('.rc'), os.listdir(prog_path)))
+def full_test_success():
+    pass
 
+def full_test_fail():
+    prog_path = os.path.join(curr_path, 'programs', 'fail')
+    prog_files = list(filter(lambda x: x.endswith('.rc'), os.listdir(prog_path)))
     outfile = 'tmp'
+    rv = 0
+    for prog in prog_files:
+        print('\nCompiling {}...'.format(prog))
+        curr_prog = os.path.join(prog_path, prog)
+        if compile_rc(curr_prog, outfile) == 0:
+            print("ERROR: program {} should have failed to compile but succeeded".format(prog))
+            rv = -1
+        else:
+            print("Program {} correctly failed to compile".format(prog))
+
+    return rv
+
+def full_test_pass():
+    
+    # another great hack to remove the dumb (program  () shit
+    def strip_eval_string(string):
+        return "'" + string[string.find('()') + 2:len(string)-2] + "'"
+    
+    prog_path = os.path.join(curr_path, 'programs', 'pass')
+    prog_files = list(filter(lambda x: x.endswith('.rc'), os.listdir(prog_path)))
+    outfile = 'tmp'
+    rv = 0
     for prog in prog_files:
         print('\nCompiling {}...'.format(prog))
         curr_prog = os.path.join(prog_path, prog)
         if compile_rc(curr_prog, outfile) != 0:
-            return dump_error('ERROR compilation failed on file {}'.format(prog), outfile)
+            rv = -1
+            continue
 
-        print('Running {}...'.format(prog))
-
-        result = subprocess.run(['./' + outfile])
-        print('\t Got result: {}'.format(result.returncode))
-
-    os.remove(outfile)
-    return 0
-
-def dump_error(msg, outfile):
-    print(msg)
-    os.remove(outfile)
-    return -1
+        # now compare the output to what the racket interpreter says
+        rc_result = subprocess.run(['./' + outfile])
         
+        with open(curr_prog, 'r') as prog_file:
+            eval_string = strip_eval_string(re.sub(' +', ' ', prog_file.read().replace('\n', ' ')))
+            print('Running racket... {}'.format(eval_string))
+            print('************************************************')
+            print('racket -e {}'.format(eval_string))
+            racket_result = subprocess.run(['racket', '-e', eval_string])
+            print(racket_result.returncode)
+
+    if os.path.exists(outfile):
+        os.remove(outfile)
+    return rv
+
+
+# compiles and runs everything in ../programs and compares the return code
+# to the output from the racket interpreter
+def full_tests():
+    full_test_fail()
+    full_test_pass()
+    return 0
+       
 def main():
     parser = argparse.ArgumentParser(description='Script to run unit/integration tests')
     parser.add_argument('-u', '--unit', action='store_true')
